Cockatrice moved to utils
