+++ packages/shared/src/utils/cockatrice/prettifyXml.ts
