+++ packages/shared/src/utils/index.ts
@@ -4,3 +4,6 @@ export * from './orderColors.ts';
 export * from './textHandling.ts';
 export * from './pipsHandling.ts';
 export * from './cardHandling';
+export * from './cockatrice';
+// export * from './draftmancer'
+// export * from './tts'
